+++ rss-client/src/app/app-routing.module.ts
@@ -6,7 +6,7 @@ import { QuizPageComponent } from './Quiz/components/quiz-page/quiz-page.compone
 import { AccountSettingsPageComponent } from './Account/components/account-settings-page/account-settings-page.component';
 import { IndividualQuizPageComponent } from './Quiz/components/individual-quiz-page/individual-quiz-page.component';
 import { LandingPageComponent } from './LandingPage/components/landing-page/landing-page.component';
-import { ShoppingCartComponent } from './shopping-cart/shopping-cart.component';
+import { ShoppingCartComponent } from './Cart/shopping-cart/shopping-cart.component';
 import { AdminviewComponent } from './inventory/adminview/adminview.component';
 import { InventoryViewComponent } from './inventory/inventory-view/inventory-view.component';
 
@@ -18,9 +18,9 @@ const routes: Routes = [
   { path: 'quizzes/:id', component: IndividualQuizPageComponent },
   { path: 'account/settings', component: AccountSettingsPageComponent },
   { path: 'cart', component: ShoppingCartComponent},
-  { path: '', component: LandingPageComponent },
   { path: 'adminview', component: AdminviewComponent },
   { path: 'inventoryview', component: InventoryViewComponent },
+  { path: '', component: LandingPageComponent },
 ];
 
 @NgModule({
